Clean up Half implementation and unit tests
